requirements.txt
appdirs==1.4.4
asgiref==3.6.0
baron==0.10.1
dj-database-url==1.2.0
dj-email-url==1.0.6
Django==4.1.6
django-cache-url==3.4.4
django-cors-headers==3.13.0
django-environs==0.1.1
django-rest-registration==0.7.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
environs==9.5.0
gunicorn==20.1.0
marshmallow==3.19.0
packaging==23.0
psycopg2==2.9.5
PyJWT==2.6.0
python-dotenv==0.21.1
pytz==2022.7.1
redbaron==0.9.2
rply==0.7.8
sqlparse==0.4.3
whitenoise==6.4.0
